requirements.txt
# requirements. 파일 내용
ultralytics
Flask
python-dotenv
pyyaml
requests
Pillow